fix: add block-timing command
