+++ applications/tari_base_node/src/commands/parser.rs
@@ -40,12 +40,9 @@ pub enum BaseNodeCommand {
     Help,
     GetPeer,
     RewindBlockchain,
-    ListHeaders,
     CheckDb,
     PeriodStats,
     HeaderStats,
-    BlockTiming,
-    CalcTiming,
     ListReorgs,
     DiscoverPeer,
     GetBlock,
